- Fixes the problem in TripalEntity::preSave that caused tripal_entity entries to be removed when only a single value was null. Now, the delta remove logic is only triggered if 1. there are values in $prop_values, and all are null. For cases where there is only a single property value per field this behavior is identical to the previos behavior. For multiple property values per field like in tripa_entity__gene_organism, the new behavior is the correct one.
- The change however brings about discrepancies rooted  deeper in the backend-code.
On the surface, this works fine for simple fields like Organism genus, species, etc.
There are difference how entries end up in complicated cases where tables are connected via linkers, eg. gene_organism.

If imported via ChadoPublish, values from organism tables are cached in tripal_entity__gene_organism. If edited via the
UI these value are not filled in. The reason is that there are two different pathways and implementatons for the same functionality:
 -- UI-TripalEntity-ChadoStorage-ChadoRecord (deep OO code)
 -- ChadoPublish (SQL heavy code)

The best way to guarantee the invocations are functionally identical would be to refactor redundant functionality into
a single method that is used by both UI and Publish.